LayoutTests/fast/dom/shadow/shadow-root-js-api.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<div id="console"></div>
<script>
description("Tests for ShadowRoot JS APIs. Can only run within DRT");

if (window.layoutTestController)
    layoutTestController.dumpAsText();
shouldThrow("new WebKitShadowRoot()");
shouldThrow("new WebKitShadowRoot(undefined)");
shouldThrow("new WebKitShadowRoot(null)");
shouldThrow("new WebKitShadowRoot(1)");
var shadowHost = document.createElement('div');
var shadowRoot = new WebKitShadowRoot(shadowHost);
shouldBe("shadowHost", "shadowRoot.host");
shouldThrow("new WebKitShadowRoot(shadowHost)");
debug('The previous test should not have side effects.');
shouldBe("shadowHost", "shadowRoot.host");

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
